Adding README

+++ ft_atoi.c
@@ -1,26 +1,26 @@
-/*************************************************************************** */
+/* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: felsanto <felsanto@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/18 10:21:31 by felsanto          #+#    #+#             */
-/*   Updated: 2022/05/19 13:32:01 by felsanto         ###   ########.fr       */
+/*   Created: 2022/06/02 16:33:57 by felsanto          #+#    #+#             */
+/*   Updated: 2022/06/02 16:40:38 by felsanto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "libft.h"
 
 int	ft_atoi(const char *str)
 {
-	unsigned int res;
-	int sign;
-	int i;
-	
+	int	res;
+	int	sign;
+	int	i;
+
 	res = 0;
 	i = 0;
 	sign = 1;
-
 	while (str[i] == ' ' || str[i] == '\t' || str[i] == '\v' \
 		|| str[i] == '\f' || str[i] == '\n' || str[i] == '\r')
 			i++;
@@ -32,9 +32,8 @@ int	ft_atoi(const char *str)
 		res = res * 10 + (str[i] - '0');
 		i++;
 	}
-	return ((int)(sign * res));
+	return (sign * res);
 }
-
 /*
 int main()
 {
